requirements.txt
tensorflow-cpu
streamlit
opencv-python-headless
numpy
Pillow
scikit-learn
matplotlib
seaborn
